requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
distro==1.9.0
fastapi==0.115.12
fastembed==0.6.0
fastembed-haystack==1.4.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.1
google-auth==2.38.0
google-genai==1.7.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
haystack-ai==2.11.2
haystack-experimental==0.8.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_imports==0.4.0
loguru==0.7.3
MarkupSafe==3.0.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
onnxruntime==1.21.0
openai==1.68.2
packaging==24.2
pillow==11.1.0
portalocker==2.10.1
posthog==3.21.0
protobuf==5.29.4
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qdrant-client==1.13.3
qdrant-haystack==9.1.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
setuptools==78.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
